refactor: libs 내부의 객체에 대해 대문자로 변경
