+++ src/components/AppHeader/Icon.ts
@@ -1,4 +1,4 @@
-import { Raw } from "../../libs/elements";
+import { Raw } from "../../libs/Elements";
 
 export const Icon = () =>
   Raw(`<svg width="24" height="25" viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
