+++ src/index.ts
@@ -1,7 +1,7 @@
 import { AppHeader } from "./components/AppHeader";
 import { Medias } from "./components/Medias";
 import { TopSection } from "./components/TopSection";
-import { Div } from "./libs/elements";
+import { Div } from "./libs/Elements";
 
 export const App = () => {
   return Div({
